requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.1.0

# Deep learning
tensorflow>=2.10.0
torch>=1.12.0
torchvision>=0.13.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.15.0

# Web framework for serving models
flask>=2.2.0
fastapi>=0.85.0
uvicorn>=0.18.0

# Model evaluation and visualization
plotly>=5.10.0
tensorboard>=2.10.0

# Image processing
opencv-python>=4.6.0
Pillow>=9.2.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0 